sysmobts: The meaning of the clock value changed from v1 to v2
In v2 the calibration value is the clock error in ppb that needs
to be compensated. Create a V2 specific implementation. Write the
clock value unconditionally as it is initialized to 0 by default
and not 0xffff.

+++ src/osmo-bts-sysmo/l1_if.h
@@ -30,7 +30,7 @@ struct femtol1_hdl {
 	struct gsm_time gsm_time;
 	uint32_t hLayer1;			/* handle to the L1 instance in the DSP */
 	uint32_t dsp_trace_f;
-	uint16_t clk_cal;
+	int clk_cal;
 	uint8_t clk_src;
 	struct llist_head wlc_list;
